drop separate super user file, add code comments
