+++ app/events/events.go
@@ -512,3 +512,16 @@ func (l *TelegramListener) transformEntities(entities []tbapi.MessageEntity) *[]
 
 	return &result
 }
+
+// SuperUser for moderators
+type SuperUser []string
+
+// IsSuper checks if username in the list of super users
+func (s SuperUser) IsSuper(userName string) bool {
+	for _, super := range s {
+		if strings.EqualFold(userName, super) || strings.EqualFold("/"+userName, super) {
+			return true
+		}
+	}
+	return false
+}
